### 2020-05-27 #2262 @whedon
[PRE REVIEW]: TorchGAN: A Flexible Framework for GAN Training and Evaluation
**Submitting author:** @avik-pal (<a href="http://orcid.org/0000-0002-3938-7375">Avik Pal</a>)
**Repository:** <a href="https://github.com/torchgan/torchgan/" target ="_blank">https://github.com/torchgan/torchgan/</a>
**Version:** v0.0.4
