### 2020-05-27 #2262 @danielskatz
👋 @avik-pal - Thanks for your submission to JOSS. As described in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss) announcing the reopening of JOSS, we're currently working in a "reduced service mode", limiting the number of papers assigned to any individual editor.

Since reopening JOSS last week, we've had > 60 papers submitted and as such, yours has been put in our backlog that we will be working through over the coming weeks and months.

Thanks in advance for your patience!

(Also, I see at least some small issues in the paper, one of the references (Krizhevesk) seems incomplete, and some things need to kept in upper case (ieee -> IEEE) - feel free to make change to the bib or md files, and enter ```@whedon generate pdf from branch ap/joss_paper``` as a new comment here to check the results)
