+++ paper/paper.bib
@@ -207,13 +207,14 @@
 @inproceedings{Krizhevsky2009LearningML,
   title={{Learning Multiple Layers of Features from Tiny Images}},
   author={Alex Krizhevsky},
-  year={2009}
+  year={2009},
+  publisher={Citeseer}
 }
 
 @InProceedings{mocogan,
 author = {Tulyakov, Sergey and Liu, Ming-Yu and Yang, Xiaodong and Kautz, Jan},
 title = {{MoCoGAN: Decomposing Motion and Content for Video Generation}},
-booktitle = {The IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
+journal = {The IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
 month = {June},
 year = {2018}
 }
