### 2020-06-12 #2262 @labarba
👋  @terrytangyuan -- Could you have a look at this submission and let us know if you can handle it as editor?
Title — TorchGAN: A Flexible Framework for GAN Training and Evaluation
[Article proof](https://github.com/openjournals/joss-papers/blob/joss.02262/joss.02262/10.21105.joss.02262.pdf)
