### 2020-10-05 #2606 @kthyng
Hello! I am the rotating associate editor in chief and am checking in on stale reviews. Will reviewers @Paul92, @urbanophile have a chance to start on/continue your reviews sometime soon? It's been over a month since the review started. Thanks!
