### 2021-04-30 #2606 @arfon
@vs74 @NMontanaBrown - would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

The submission we'd be asking you to review is [_TorchGAN: A Flexible Framework for GAN Training and Evaluation_](https://github.com/torchgan/torchgan/)
