### 2021-08-09 #2606 @NMontanaBrown
> 👋 folks. Checking in here to see how things are going. @NMontanaBrown it seems like you made it a good way through your review but might be waiting on @avik-pal to make some updates?

I am waiting on the updates regarding unit tests to assert the functionality of the package, raised in this [ticket](https://github.com/torchgan/torchgan/issues/138) before finishing the review.
