### 2021-09-16 #2606 @arfon
:wave: @yxoos –  would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

We already have one review from @NMontanaBrown and are looking for a second reviewer to help us wrap up the JOSS review process.

The submission we'd be asking you to review is [_TorchGAN: A Flexible Framework for GAN Training and Evaluation_](https://github.com/torchgan/torchgan/). We already have one review complete and are seeking a second reviewer to wrap up our review process.
