### 2021-10-05 #2606 @terrytangyuan
@avik-pal Just some small feedback:
1. There might be some papers that you cited in arXiv but have been published to somewhere more formally. Could you check and update them accordingly?
2. The paper has been here for a year and some things might have been changed, such as the feature tables for existing frameworks. Could you update those wherever necessary? If the API of TorchGAN has changed, please update as well.
