### 2021-10-16 #2606 @avik-pal
Yes the latest pdf has the updated references
